requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10
schedule>=1.2.1
python-dotenv>=1.0.0
jinja2>=3.1.2
